requirements.txt

# Additional requirements go here
django-rest-framework